+++ kong/runloop/plugin_servers/init.lua
@@ -4,6 +4,8 @@ local cjson = require "cjson.safe"
 local clone = require "table.clone"
 local ngx_ssl = require "ngx.ssl"
 local SIGTERM = 15
+local type = type
+local pairs = pairs
 
 local type = type
 local pairs = pairs
@@ -109,6 +111,12 @@ local exposed_api = {
     end
 
     return header_value
+  end,
+
+  ["kong.request.get_uri_captures"] = function ()
+    local saved = get_saved()
+    local ngx_ctx = saved and saved.ngx_ctx or ngx.ctx
+    return kong.request.get_uri_captures(ngx_ctx)
   end,
 
   ["kong.response.get_status"] = function()