+++ CHANGELOG.md
@@ -109,6 +109,10 @@
 - Increase the maximum request argument number from `100` to `1000`, and return 400 error if request parameters reach the limitation to avoid being truncated.
   [#9510](https://github.com/Kong/kong/pull/9510)
 
+#### PDK
+
+- Added support for `kong.request.get_uri_captures`(`kong.request.getUriCaptures`)
+  [#9512](https://github.com/Kong/kong/pull/9512)
 
 #### Plugins